+++ LogParser/DriverCall/Functions/Map.cs
@@ -8,17 +8,17 @@ namespace Migoto.Log.Parser.DriverCall
         {
         }
 
-        public uint pResource { get; set; }
+        public ulong pResource { get; set; }
         public uint Subresource { get; set; }
         public uint MapType { get; set; }
         public uint MapFlags { get; set; }
-        public uint pMappedResource { get; set; }
+        public ulong pMappedResource { get; set; }
 
         public Asset.Base Resource { get; set; }
 
         Asset.Base IResource.Asset => Resource;
         int IResource.Index => -1;
-        uint IResource.Pointer => pResource;
+        ulong IResource.Pointer => pResource;
         Base IResource.Owner => this;
 
         public void UpdateAsset(Asset.Base asset) => Resource = asset;
