+++ LogParser/DriverCall/Functions/Unmap.cs
@@ -8,14 +8,14 @@ namespace Migoto.Log.Parser.DriverCall
         {
         }
 
-        public uint pResource { get; set; }
+        public ulong pResource { get; set; }
         public uint Subresource { get; set; }
 
         public Asset.Base Resource { get; set; }
 
         Asset.Base IResource.Asset => Resource;
         int IResource.Index => -1;
-        uint IResource.Pointer => pResource;
+        ulong IResource.Pointer => pResource;
         Base IResource.Owner => this;
 
         public void UpdateAsset(Asset.Base asset) => Resource = asset;
