+++ LogParser/DriverCall/Shader/SetConstantBuffers.cs
@@ -14,7 +14,7 @@ namespace Migoto.Log.Parser.DriverCall
 
         public uint NumBuffers { get; set; }
 
-        public uint ppConstantBuffers { get; set; }
+        public ulong ppConstantBuffers { get; set; }
 
         public List<Resource> ConstantBuffers { get; set; } = new List<Resource>(14);
     }
