+++ LogParser/DriverCall/Shader/SetSamplers.cs
@@ -14,7 +14,7 @@ namespace Migoto.Log.Parser.DriverCall
 
         public uint NumSamplers { get; set; }
 
-        public uint ppSamplers { get; set; }
+        public ulong ppSamplers { get; set; }
 
         public List<Sampler> Samplers { get; set; } = new List<Sampler>(16);
     }
